Adding delete methods
