+++ src/routes/main.js
@@ -1,5 +1,6 @@
 import { setGetForm, setObjectView } from "./get/get.js";
 import { setPostForm, setPostView } from "./post/post.js";
+import { setDeleteDescription, setDeleteForm, setDeletedView } from "./delete/delete.js";
 
 let isResumeDeleted = false;
 let isStylingDeleted = false;
@@ -74,74 +75,74 @@ function setPatchForm() {}
 
 function setPutForm() {}
 
-function setDeleteForm() {
-  // resetForm(document.getElementById("demo"));
-  document.getElementById("demo").innerHTML = "...personal-site/resource?type=";
-  var form = document.createElement("form");
-  form.id = "delete-object";
-  select = document.createElement("select");
-  select.id = "object-choice";
-  form.appendChild(select);
+// function setDeleteForm() {
+//   // resetForm(document.getElementById("demo"));
+//   document.getElementById("demo").innerHTML = "...personal-site/resource?type=";
+//   var form = document.createElement("form");
+//   form.id = "delete-object";
+//   select = document.createElement("select");
+//   select.id = "object-choice";
+//   form.appendChild(select);
 
-  optionDefault = document.createElement("option");
-  optionDefault.value = "";
-  optionDefault.hidden = true;
-  optionDefault.disabled = true;
-  optionDefault.selected = true;
-  select.appendChild(optionDefault);
+//   optionDefault = document.createElement("option");
+//   optionDefault.value = "";
+//   optionDefault.hidden = true;
+//   optionDefault.disabled = true;
+//   optionDefault.selected = true;
+//   select.appendChild(optionDefault);
 
-  option1 = document.createElement("option");
-  option1.value = "resume";
-  option1.innerHTML = "Resume";
-  select.appendChild(option1);
+//   option1 = document.createElement("option");
+//   option1.value = "resume";
+//   option1.innerHTML = "Resume";
+//   select.appendChild(option1);
 
-  option2 = document.createElement("option");
-  option2.value = "styling";
-  option2.innerHTML = "Styling";
-  select.appendChild(option2);
+//   option2 = document.createElement("option");
+//   option2.value = "styling";
+//   option2.innerHTML = "Styling";
+//   select.appendChild(option2);
 
-  option3 = document.createElement("option");
-  option3.value = "everything";
-  option3.innerHTML = "Everything";
-  select.appendChild(option3);
+//   option3 = document.createElement("option");
+//   option3.value = "everything";
+//   option3.innerHTML = "Everything";
+//   select.appendChild(option3);
 
-  document.getElementById("demo").after(form);
-}
+//   document.getElementById("demo").after(form);
+// }
 
-function setDeleteDescription(object) {
-  document.getElementById("render-object").innerHTML = "";
-  switch (object) {
-    case "resume":
-      document.getElementById("render-object").innerHTML = "resumedeleting";
-      break;
-    case "styling":
-      document.getElementById("render-object").innerHTML = "stylgingdeleting";
-      break;
-    case "everything":
-      document.getElementById("render-object").innerHTML = "everythingdeleting";
-      break;
-  }
-}
+// function setDeleteDescription(object) {
+//   document.getElementById("render-object").innerHTML = "";
+//   switch (object) {
+//     case "resume":
+//       document.getElementById("render-object").innerHTML = "resumedeleting";
+//       break;
+//     case "styling":
+//       document.getElementById("render-object").innerHTML = "stylgingdeleting";
+//       break;
+//     case "everything":
+//       document.getElementById("render-object").innerHTML = "everythingdeleting";
+//       break;
+//   }
+// }
 
-function setDeletedView(object) {
-  switch (object) {
-    case "resume":
-      isResumeDeleted = true;
-      document.getElementById("render-object").innerHTML =
-        "Welp, you deleted the resume, damn!";
-      break;
-    case "styling":
-      isStylingDeleted = true;
-      document.styleSheets[0].disabled = !document.styleSheets[0].disabled;
-      document.getElementById("render-object").innerHTML =
-        "Welp, you deleted all the styling on this website!";
-      break;
-    case "everything":
-      document.getElementsByTagName("html")[0].innerHTML =
-        "Welp, you deleted everything on this website! Try reloading to reset";
-      break;
-  }
-}
+// function setDeletedView(object) {
+//   switch (object) {
+//     case "resume":
+//       isResumeDeleted = true;
+//       document.getElementById("render-object").innerHTML =
+//         "Welp, you deleted the resume, damn!";
+//       break;
+//     case "styling":
+//       isStylingDeleted = true;
+//       document.styleSheets[0].disabled = !document.styleSheets[0].disabled;
+//       document.getElementById("render-object").innerHTML =
+//         "Welp, you deleted all the styling on this website!";
+//       break;
+//     case "everything":
+//       document.getElementsByTagName("html")[0].innerHTML =
+//         "Welp, you deleted everything on this website! Try reloading to reset";
+//       break;
+//   }
+// }
 
 //The file is a lot of things, maybe split up the file to perform different things?
 // JSX is better? (do research on that)
